requirements.txt
async-timeout==4.0.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
et-xmlfile==1.1.0
google-auth==2.21.0
google-auth-oauthlib==1.0.0
googlesearch-python==1.2.3
gspread==5.10.0
idna==3.4
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.2
Pillow==9.5.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
redis==4.6.0
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.5
spotipy==2.23.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
Wikipedia-API==0.6.0
